Refactor imports to point to ~/graphql/data/queries for better organization and clarity across utility files and routes
